@@ -9,6 +9,7 @@ from pycardano import PlutusData
 from pycardano import PlutusV1Script
 from pycardano import PlutusV2Script
 from pycardano import Redeemer
+from pycardano import TransactionBuilder
 from pycardano import TransactionOutput
 from pycardano import UTxO
 
@@ -120,7 +121,6 @@ class AbstractPairState(DendriteBaseModel, ABC):
         """
         return PlutusV1Script
 
-    @property
     def script_class(self) -> type[PlutusV1Script] | type[PlutusV2Script]:
         """Returns the script class based on the Plutus version being used."""
         if self.plutus_v2:
@@ -145,10 +145,11 @@ class AbstractPairState(DendriteBaseModel, ABC):
         address_source: Address,
         in_assets: Assets,
         out_assets: Assets,
+        tx_builder: TransactionBuilder | None = None,
         extra_assets: Assets | None = None,
         address_target: Address | None = None,
         datum_target: PlutusData | None = None,
-    ) -> TransactionOutput:
+    ) -> tuple[TransactionOutput | None, PlutusData]:
         """Constructs the transaction output for a swap."""
         raise NotImplementedError
 
